fix(nimble): Fix the size of num_packets in dtm event

+++ nimble/host/include/host/ble_gap.h
@@ -1174,7 +1174,7 @@ struct ble_gap_event {
 	     * Valid only for BLE_GAP_DTM_END_EVT
 	     * shall be 0 for a transmitter.
 	     */
-            uint8_t num_pkt;
+            uint16_t num_pkt;
 	} dtm_state;
     };
 };
